refactor: move utility functions to separate module and update tsconfig settings
